Enable arch build by suppressing isl library
Arch, being a rolling release has a much newer version of libisl which
breaks the gcc buidl unless we suppress the gcc build's use of isl,
which is entirely do-able.
There's a slight issue in that this isn't selectable from the command
line or via any config mechanism, so hide it away inside a branch.